fix(lots): avoid deleting when no deliveries exist

+++ routes/delivery_lots.py
@@ -463,9 +463,10 @@ async def delivery_lots_id_plan_get(
             pth_db = db.get(DeliveryPath, route.route_id)
 
             # Loop old points and remove old relations
-            for link in pth_db.deliveries:
-                db.delete(link)
-                db.commit()
+            if pth_db.deliveries:
+                for link in pth_db.deliveries:
+                    db.delete(link)
+                    db.commit()
 
             # Loop new points and create new relations
             for waypoint in route.optimized_waypoints:
